third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-number-space-list-expected.txt

This is a testharness.js-based test.
[PASS] Animating a custom property of type <number>+
[PASS] Animating a custom property of type <number>+ with a single keyframe
[PASS] Animating a custom property of type <number>+ with additivity
[PASS] Animating a custom property of type <number>+ with a single keyframe and additivity
[FAIL] Animating a custom property of type <number>+ with iterationComposite
  assert_equals: expected "250 275" but got "50 75"
[PASS] Animating a custom property of type <number>+ with different lengths is discrete
Harness: the test ran to completion.
